+++ consensus/merge/src/test/java/org/hyperledger/besu/consensus/merge/blockcreation/MergeCoordinatorTest.java
@@ -183,7 +183,8 @@ public class MergeCoordinatorTest implements MergeGenesisConfigHelper {
         .when(protocolSchedule)
         .getByBlockHeader(any(BlockHeader.class));
 
-    protocolContext = new ProtocolContext(blockchain, worldStateArchive, mergeContext);
+    protocolContext =
+        new ProtocolContext(blockchain, worldStateArchive, mergeContext, Optional.empty());
     var mutable = worldStateArchive.getMutable();
     genesisState.writeStateTo(mutable);
     mutable.persist(null);
@@ -255,7 +256,6 @@ public class MergeCoordinatorTest implements MergeGenesisConfigHelper {
                       protocolSchedule,
                       this.miningParameters.getMinTransactionGasPrice(),
                       address.or(miningParameters::getCoinbase).orElse(Address.ZERO),
-                      this.miningParameters.getMinBlockOccupancyRatio(),
                       parentHeader,
                       Optional.empty()));