+++ plugins/rocksdb/src/main/java/org/hyperledger/besu/plugin/services/storage/rocksdb/segmented/RocksDBColumnarKeyValueStorage.java
@@ -87,7 +87,9 @@ public abstract class RocksDBColumnarKeyValueStorage
     RocksDbUtil.loadNativeLibrary();
   }
 
-  private final AtomicBoolean closed = new AtomicBoolean(false);
+  /** atomic boolean to track if the storage is closed */
+  protected final AtomicBoolean closed = new AtomicBoolean(false);
+
   private final WriteOptions tryDeleteOptions =
       new WriteOptions().setNoSlowdown(true).setIgnoreMissingColumnFamilies(true);
   private final ReadOptions readOptions = new ReadOptions().setVerifyChecksums(false);
@@ -326,6 +328,11 @@ public abstract class RocksDBColumnarKeyValueStorage
           .forEach(ColumnFamilyHandle::close);
       getDB().close();
     }
+  }
+
+  @Override
+  public boolean isClosed() {
+    return closed.get();
   }
 
   void throwIfClosed() {